refactor: moved the ticket components into its own feature dir
