+++ src/features/ticket/components/ticketCard.tsx
@@ -20,7 +20,6 @@ export default function TicketCard(props: TripCardProps) {
 
         <div className="flex flex-row justify-between items-center">
           <div className="flex flex-row gap-1">
-            {" "}
             {/*Left Side*/}
             <button>
               <User />
@@ -28,7 +27,6 @@ export default function TicketCard(props: TripCardProps) {
             <span className="text-[#525252] text-lg">{props.passenger}</span>
           </div>
           <div className="flex flex-row items-center gap-2">
-            {" "}
             {/*Right Side*/}
             <SquarePen />
             <Button className="h-[70%] rounded-lg bg-[#B81F1F] hover:bg-[#8B1919] font-semibold text-lg">
